requirements.txt
httpx~=0.27.0
python-dotenv~=1.0.1
langchain~=0.2.10